Clean up MultiWorkerExecutor and move WorkerSynchronizer to separate file
